requirements.txt
streamlit
transformers
torch
moviepy
pillow